New backup script which will mount a /backup, then use rsync to copy files from /usr/home to /backup/yyyy-mm-dd. After it is done, it will remove the oldest backup, 14 days ago. All paths are hardcoded.
If a file .exclude is in a users homedir, it is looked at, and is used to exclude certain directories or files from the rsync. For example:

in /usr/home/xistence/

public_html/iso

No trailing slash, this will cause the folder /usr/home/xistence/public_html/iso to not be backed up. This might be for various reasons, but for this example it is because backing up GB sized files should not be something that is done, only because it takes ages and wastes valuable space. 

Thing to keep in mind is, if you want to backup a 40 GB HD, with 22 GB used, to keep even a 3 day backup of that would require 3 * 22 GB, or 66 GB of available space on the backup drive. It is therefore in ones best interest to exclude very large files, and back them up in a different method since they are most likely not going to change much across revisions.

git-svn-id: http://svn.0x58.com/repos/bsdpanel/trunk@206 db258d53-41f1-0310-98c4-cc461a13e28c